+++ environment.yml
@@ -1,0 +1,13 @@
+name: gsp_api
+channels:
+  - conda-forge
+dependencies:
+  - python=3.7
+  - uwsgi
+  - flask
+  - flask-restful
+  - flask-cors
+  - pandas
+  - xarray
+  - netCDF4
+  - shapely
