+++ supervisord.conf
@@ -3,7 +3,7 @@ nodaemon=true
 
 [program:uwsgi]
 directory=/app/GSP_API/
-command=/usr/local/envs/gsp_api/bin/uwsgi --virtualenv /usr/local/envs/gsp_api --callable app --http 0.0.0.0:80 -b 32768 --wsgi-disable-file-wrapper --die-on-term --enable-threads --wsgi-file /app/GSP_API/runserver.py --log-date="%%Y-%%m-%%d %%H:%%M:%%S" --logformat-strftime
+command=/opt/conda/envs/gsp_api/bin/uwsgi --virtualenv /opt/conda/envs/gsp_api --callable app --http 0.0.0.0:80 -b 32768 --wsgi-disable-file-wrapper --die-on-term --enable-threads --wsgi-file /app/GSP_API/runserver.py --log-date="%%Y-%%m-%%d %%H:%%M:%%S" --logformat-strftime
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
 stderr_logfile=/dev/stdout
